Merges foreign key constraints to allow non-unique room numbers
Originally each room number was forced to be unique. Now it is only the
combination of chain, hotel, and room number that has to be unique. The
foreign key constraints that depended on this uniqueness have also been
updated. This operation also needs to be performed for hotels, but has
not yet been done.